refactor: rename errors/memory.py to errors/memory_store.py
