+++ src/llm_agents_from_scratch/errors/__init__.py
@@ -5,7 +5,11 @@ from .core import (
     MissingExtraError,
 )
 from .mcp import MCPError, MCPWarning, MissingMCPServerParamsError
-from .memory import EpisodeNotFoundWarning, MemoryStoreError, MemoryStoreWarning
+from .memory_store import (
+    EpisodeNotFoundWarning,
+    MemoryStoreError,
+    MemoryStoreWarning,
+)
 from .skill import (
     EmptySkillBodyError,
     InvalidFrontmatterError,
